Any trial & errors not working were moved to "NO" versions
