+++ src/main/java/frc/robot/commands/Launcher/SemiAutoVersion7_NO.java
@@ -1,0 +1,24 @@
+// Copyright (c) FIRST and other WPILib contributors.
+// Open Source Software; you can modify and/or share it under the terms of
+// the WPILib BSD license file in the root directory of this project.
+
+
+package frc.robot.commands.Launcher;
+import edu.wpi.first.wpilibj2.command.CommandBase;
+import frc.robot.commands.IndexSpeed;
+import frc.robot.subsystems.Index;
+import frc.robot.subsystems.Launcher;
+
+
+public class SemiAutoVersion7_NO {
+    public void SemiAutoVersion3(Index indexMotors, Launcher launcher){
+    // double tx = NetworkTableInstance.getDefault().getTable("limelight-back").getEntry("tx").getDouble(0);
+    double tx = 2; // Defining for testing purposes
+    // double ty = NetworkTableInstance.getDefault().getTable("limelight-back").getEntry("ty").getDouble(0);
+    
+      if (tx < 5) {
+          new IndexSpeed(indexMotors, 0.5);
+         } else {
+          new LaunchSpeed(launcher, 0.5, 0.5);}
+    }
+} // end of class
